+++ src/app/_components/product-overview/product-overview.component.ts
@@ -1,22 +1,29 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { Product } from '@interfaces/productDto';
 import { Router } from '@angular/router';
 import { WishlistService } from '@services/wishlist.service';
 import { CartService } from '@services/cart.service';
+import { TranslateService } from '@ngx-translate/core';
+import { DialogOptions } from '@interfaces/dialogOptionsDto';
+import { ConfirmDialogService } from '@services/confirm-dialog.service';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-product-overview',
   templateUrl: './product-overview.component.html',
   styleUrls: ['./product-overview.component.scss'],
 })
-export class ProductOverviewComponent implements OnInit {
+export class ProductOverviewComponent implements OnInit, OnDestroy {
   @Input() product!: Product;
   isHovered = false;
+  private removeSubscription$: Subscription | undefined;
 
   constructor(
     private router: Router,
     private wishlistService: WishlistService,
-    private cartService: CartService
+    private cartService: CartService,
+    private translate: TranslateService,
+    private dialogService: ConfirmDialogService
   ) {}
 
   goToProductDetails() {
@@ -29,8 +36,20 @@ export class ProductOverviewComponent implements OnInit {
   }
 
   removeFromCart() {
-    this.cartService.removeFromCart(this.product);
-    this.product.isInCart = false;
+    const confirmOptions: DialogOptions = {
+      title: this.translate.instant('common.areYouSure'),
+      message: this.translate.instant('products.confirmDeleteFromCart'),
+    };
+    this.dialogService.open(confirmOptions);
+
+    this.removeSubscription$ = this.dialogService
+      .confirmed()
+      .subscribe((confirmed) => {
+        if (confirmed) {
+          this.cartService.removeFromCart(this.product);
+          this.product.isInCart = false;
+        }
+      });
   }
 
   addToWishlist() {
@@ -46,5 +65,9 @@ export class ProductOverviewComponent implements OnInit {
   ngOnInit() {
     this.wishlistService.manageProduct(this.product);
     this.cartService.manageProduct(this.product);
+  }
+
+  ngOnDestroy(): void {
+    if (this.removeSubscription$) this.removeSubscription$.unsubscribe();
   }
 }
